requirements.txt
pymodbus==2.5.3
pyqtgraph==0.13.7
pyserial==3.5
PyQt5==5.15.11
